+++ R/manyenviron-agreements.R
@@ -6,20 +6,22 @@
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{IEADB: }{A dataset with 3666 observations and the following
-#' 10 variables: manyID, Title, Beg, DocType, AgreementType, Signature, Force, Lineage, treatyID, ieadbID.}
-#' \item{GNEVAR: }{A dataset with 3428 observations and the following
-#' 12 variables: manyID, Title, Beg, End, DocType, AgreementType, GeogArea, Signature, Force, Lineage, treatyID, gnevarID.}
+#' 10 variables: manyID, Title, Beg, DocType, AgreementType, Signature,
+#' Force, Lineage, treatyID, ieadbID.}
+#' \item{GNEVAR: }{A dataset with 7273 observations and the following
+#' 14 variables: manyID, Title, Beg, End, DocType, AgreementType, GeogArea,
+#' Signature, Force, Lineage, accessionC, accessionP, treatyID, gnevarID.}
 #' \item{ECOLEX: }{A dataset with 2174 observations and the following
-#' 10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force, Lineage, treatyID, ecolexID.}
+#' 10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force,
+#' Lineage, treatyID, ecolexID.}
 #' \item{CIESIN: }{A dataset with 666 observations and the following
 #' 7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
 #' \item{HEIDI: }{A dataset with 2280 observations and the following
 #' 7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
 #' }
-
 #'
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}
-#' lapply(agreements, skimr::skim_without_charts)
+#' lapply(agreements, messydates::mreport)
 #' ```
 "agreements"