+++ R/manyenviron-texts.R
@@ -1,17 +1,18 @@
 #' texts database documentation
 #'
+#' @import manydata
 #' @format The texts database is a list that contains the
-#' following 1 datasets: AGR_TXT.
+#' following 1 datasets: GNEVAR_TXT.
 #' For more information and references to each of the datasets used,
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
-#' \item{AGR_TXT: }{A dataset with 6377 observations and the following
-#' 8 variables: manyID, Title, Beg, Text, Source, ieadbID, gnevarID, ecolexID.}
+#' \item{GNEVAR_TXT: }{A dataset with 6377 observations and the following
+#' 8 variables: manyID, Title, Beg, TreatyText, Source, ieadbID, gnevarID,
+#' ecolexID.}
 #' }
-
-#'
+#' 
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}
-#' lapply(texts, skimr::skim_without_charts)
+#' lapply(texts, messydates::mreport)
 #' ```
 "texts"