+++ data-raw/memberships/ECOLEX_MEM/prepare-ECOLEX.R
@@ -15,10 +15,10 @@ manypkgs::retain("ECOLEX_MEM")
 ECOLEX_MEM <- as_tibble(ECOLEX_MEM) %>%
   dplyr::rename(For = Force,
                 Rati = Rat) %>%
-  manydata::transmutate(SignatureCountry = manypkgs::standardise_dates(Sign),
-                     End = manypkgs::standardise_dates(Term),
-                     Force = manypkgs::standardise_dates(For),
-                     Rat = manypkgs::standardise_dates(Rati),
+  manydata::transmutate(SignatureCountry = messydates::as_messydate(Sign),
+                     End = messydates::as_messydate(Term),
+                     Force = messydates::as_messydate(For),
+                     Rat = messydates::as_messydate(Rati),
                      CountryID = StatID,
                      ecolexID = EcolexID) %>%
   dplyr::mutate(Beg = dplyr::coalesce(SignatureCountry, Rat, Force)) %>%