+++ data-raw/regimes/IRD/prepare-IRD.R
@@ -11,13 +11,14 @@ IRD <- readxl::read_excel("data-raw/regimes/IRD/IRD.xlsx")
 # formats of the 'IRD' object until the object created
 # below (in stage three) passes all the tests.
 IRD <- as_tibble(IRD) %>%
-  manydata::transmutate(Beg = manypkgs::standardise_dates(Formation),
-                        End = manypkgs::standardise_dates(as.character(Endpoint)),
+  manydata::transmutate(Beg = messydates::as_messydate(Formation),
+                        End = messydates::as_messydate(as.character(Endpoint)),
                         # Watershed and Watershed2 refer to periods of fundamental change in the regime
-                        Wat = manypkgs::standardise_dates(Watershed),
-                        Wat2 = manypkgs::standardise_dates(Watershed2)) %>%
-  dplyr::arrange(Beg) %>% 
-  # RegimeComponent is an institutional arrangement that is part of a given regime, such as a protocol
+                        Wat = messydates::as_messydate(Watershed),
+                        Wat2 = messydates::as_messydate(Watershed2)) %>%
+  dplyr::arrange(Beg) %>%
+  # RegimeComponent is an institutional arrangement that is part of a given regime,
+  # such as a protocol
   # RegimeElement refers to a regulatory area of a given regime for a given time period
   dplyr::select(Regime, RegimeComponent, Beg, Wat, Wat2, End, RegimeElement)
 # manypkgs includes several functions that should help cleaning
@@ -25,8 +26,7 @@ IRD <- as_tibble(IRD) %>%
 # Please see the vignettes or website for more details.
 
 # Stage three: Connecting data
-# Next run the following line to make IRD available
-# within the package.
+# Next run the following line to make IRD available within the package.
 # This function also does two additional things.
 # First, it creates a set of tests for this object to ensure adherence
 # to certain standards.You can hit Cmd-Shift-T (Mac) or Ctrl-Shift-T (Windows)