+++ NEWS.md
@@ -1,3 +1,16 @@
+# manyenviron 0.1.3
+
+2022-08-16
+
+## Package
+* Closed #57 by switching back to original logo size.
+* Updated workflow files to include package caching.
+
+## Data
+* Changed class from 'messydt' to 'mdate' for date variables (`Beg`, `Signature`, `Force`, `End`, `Rat`, `SignatureCountry`, and `Term`) across all databases using `messydates::as_messydate()`.
+* Re-exported treaty texts after standardising with `manypkgs::standardise_treaty_text()`.
+* Added `accessionC` and `accessionP` variables in `GNEVAR` dataset in `agreements` database using `manypkgs::code_accession_terms()`. These variables code the accession conditions (such as whether it is open to all states) and procedures (for example by majority voting) respectively that are stipulated in treaties for states acceding to a treaty.
+
 # manyenviron 0.1.2
 
 2022-03-16
@@ -10,7 +23,6 @@
 * Fixed #55 by renaming variables in database to avoid ambiguity
   * Renamed variables "L", "D", and "J" in the agreements database "DocType", "AgreementType", and "GeogArea" respectively
   * Renamed variables "L" and "SignatureC" in memberships database "DocType" and "SignatureCountry" respectively
-
 
 # manyenviron 0.1.1